Minor improvements to playlist handling
